jit.off() shows no mem leaks

+++ examples/echo_svr_pull_style.c
@@ -47,7 +47,7 @@ int main(int argc, char *argv[]) {
                mnet_chann_socket_addr(msg->r, &addr);
                printf("svr accept cnt with chann %s:%d\n", addr.ip, addr.port);
             } else if (msg->event == CHANN_EVENT_TIMER) {
-               printf("svr current time: %zd\n", mnet_current());
+               printf("svr current time: %lld\n", (int64_t)mnet_current());
             }
             continue;
          }
